WebBackend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
filelock==3.18.0
mysqlclient==2.2.7
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pipenv==2025.0.2
platformdirs==4.3.8
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
xlrd==1.2.0
python-decouple==3.6